+++ CHANGELOG.md
@@ -27,7 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PeriodicalSurface: handles exceptions in connect_contours method.
 - ExtrusionSurface3D: fullarcellipse3d_to_2d
 - ExtrusionSurface3D: generalization of the _repair_points_order method to repair the order of parametric points of edges after transformation.
-- ToroidalSurface3D: increases precision of point3d_to_2d.
+- ToroidalSurface3D: 
+
+#### wires.py
+- ContourMixin: is_ordered()
 
 ### Refactor
 - Big refactor to improve and simplify complex and long methods in various modules.